commit de atualização

+++ codigo/src/br/com/ProjectAP/model/Estacionamento.java
@@ -15,8 +15,8 @@ public class Estacionamento {
 	private String complemento;
 	private String telefone;
 	private double valorDaVaga;
-	private long latitude;
-	private long longitude;
+	private float latitude;
+	private float longitude;
 	
 	
 	
@@ -70,16 +70,16 @@ public class Estacionamento {
 	public void setValorDaVaga(double valorDaVaga) {
 		this.valorDaVaga = valorDaVaga;
 	}
-	public long getLatitude() {
+	public float getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(float latitude) {
 		this.latitude = latitude;
 	}
-	public long getLongitude() {
+	public float getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(float longitude) {
 		this.longitude = longitude;
 	}
